Provide details on channels for users.admin.invite
Based on personal experimentation with the calls, providing a list of
channel names such as "general,random" is not allowed by the backend
and instead a list of channel IDs must be provided - which can be
requested, for example, by walking through the data returned by the
public channels.list method. Additionally, passing in names or any
other faulty details instead of existing channel IDs results in a
friendly error message saying the channel was not found. These details
may be super helpful for anyone passing by later down the line, since
having people auto-join specific channels can be very powerful, knowing
that one of the biggest strengths of Slack lies in the categorization
powers provided by channels.